moved BuildMode
